# dennic_api_gateway/api/docs/swagger.yaml
definitions:
...
  model_booking_service.CreatePatientReq:
    properties:
      address:
        type: string
      birth_date:
        type: string
      blood_group:
        type: string
      city:
        type: string
      country:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      patient_problem:
        type: string
      phone_number:
        type: string
    type: object
...
  model_booking_service.Patient:
    properties:
      address:
        type: string
      birth_date:
        type: string
      blood_group:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      patient_problem:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
  model_booking_service.PatientsType:
    properties:
      count:
        type: integer
      patients:
        items:
          $ref: '#/definitions/model_booking_service.Patient'
        type: array
    type: object
...
  model_booking_service.UpdatePatientReq:
    properties:
      address:
        type: string
      birth_date:
        type: string
      blood_group:
        type: string
      city:
        type: string
      country:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      patient_id:
        type: string
      patient_problem:
        type: string
      phone_number:
        type: string
    type: object
  model_common.ResponseError:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  model_common.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/model_common.ResponseError'
    type: object
...
  models.StatusRes:
    properties:
      status:
        type: boolean
    type: object
host: dennic.uz:9050
info:
  contact: {}
  title: Dennic Project
  version: "1.7"
paths:
...
  /v1/patient:
    delete:
      consumes:
      - application/json
      description: DeletePatient - Api for delete patient
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: DeletePatient
      tags:
      - Patient
    get:
      consumes:
      - application/json
      description: ListPatient - Api for list patient
      parameters:
      - description: searchField
        enum:
        - first_name
        - last_name
        - blood_group
        - phone_number
        - address
        - city
        - country
        in: query
        name: searchField
        type: string
      - in: query
        name: limit
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: string
      - in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.PatientsType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: ListPatient
      tags:
      - Patient
    post:
      consumes:
      - application/json
      description: CreatePatient - Api for crete patient
      parameters:
      - description: CreatePatientReq
        in: body
        name: CreatePatientReq
        required: true
        schema:
          $ref: '#/definitions/model_booking_service.CreatePatientReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: CreatePatient
      tags:
      - Patient
    put:
      consumes:
      - application/json
      description: UpdatePatient - Api for update patient
      parameters:
      - description: UpdatePatientReq
        in: body
        name: UpdatePatientReq
        required: true
        schema:
          $ref: '#/definitions/model_booking_service.UpdatePatientReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: UpdatePatient
      tags:
      - Patient
  /v1/patient/get:
    get:
      consumes:
      - application/json
      description: GetPatient - Api for get patient
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: GetPatient
      tags:
      - Patient
  /v1/patient/phone:
    put:
      consumes:
      - application/json
      description: UpdatePhonePatient - Api for update phone patient
      parameters:
      - description: phone_number
        in: query
        name: phone_number
        required: true
        type: string
      - description: new_phone_number
        in: query
        name: new_phone_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: UpdatePhonePatient
      tags:
      - Patient
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"